Clone the `bookmark` app to `alert`
We are preparing a big refactoring. The "bookmark" feature will
disappear and become an "alert" feature.

It would be complicated renaming the app, so I'm cloning it to ease the
transition.

Once the new "alert" feature is live and the existing data is migrated,
the `bookmark` app will be deleted.